fix: retrieve an appropriate unix socket path
We have assumed that we can obtain `lsof -aUc launchd_sim` result in a specific order. However, starting by Xcode 13.4, not starting by Xcode13.1, the asumption seems to be wrong. Therefore, it is most likely necessary to obtain an appropriate unix socket path without depending on the order of the output.